create weather context
