+++ frontend/components/contexts/WeatherContext.js
@@ -1,0 +1,8 @@
+import { createContext, useContext } from 'react';
+
+export const WeatherContext = createContext();
+
+export function useWeather() {
+  const weather = useContext(WeatherContext);
+  return weather;
+}
