+++ frontend/pages/_app.tsx
@@ -1,11 +1,12 @@
-import '../styles/globals.css'
-import type { AppProps } from 'next/app'
-import Footer from '../components/Footer'
+import '../styles/globals.css';
+import type { AppProps } from 'next/app';
+import Footer from '../components/Footer';
 
 function MyApp({ Component, pageProps }: AppProps) {
-  return <>
-    <Component {...pageProps} />
-    <Footer/>
-  </>
+  return (
+    <>
+      <Component {...pageProps} />
+    </>
+  );
 }
-export default MyApp
+export default MyApp;
